requirements.txt
kafka-python==2.0.2
web3==6.11.3